add installers classes
