+++ backend/InsuranceApp/src/InsuranceApp.WebApi/Installers/IInstaller.cs
@@ -1,0 +1,10 @@
+﻿using Microsoft.Extensions.Configuration;
+using Microsoft.Extensions.DependencyInjection;
+
+namespace InsuranceApp.WebApi.Installers
+{
+    public interface IInstaller
+    {
+        void InstallServices(IServiceCollection services, IConfiguration Configuration);
+    }
+}
