+++ backend/InsuranceApp/src/InsuranceApp.WebApi/Installers/InstallerExtensions.cs
@@ -1,0 +1,18 @@
+﻿using System;
+using System.Linq;
+using Microsoft.Extensions.Configuration;
+using Microsoft.Extensions.DependencyInjection;
+
+namespace InsuranceApp.WebApi.Installers
+{
+    public static class InstallerExtensions
+    {
+        public static void InstallServicesInAssembly(this IServiceCollection services, IConfiguration configuration)
+        {
+            var installers = typeof(Startup).Assembly.ExportedTypes.Where(x =>
+                typeof(IInstaller).IsAssignableFrom(x) && !x.IsInterface && !x.IsAbstract).Select(Activator.CreateInstance).Cast<IInstaller>().ToList();
+
+            installers.ForEach(installer => installer.InstallServices(services, configuration));
+        }
+    }
+}
